data-pipeline/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
playwright==1.40.0
lxml==4.9.3
fake-useragent==1.4.0

# Data Processing & ETL
pandas==2.1.3
numpy==1.25.2
pyarrow==14.0.1
polars==0.19.12
dask==2023.10.1

# Database & Storage
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
pymongo==4.6.0
redis==5.0.1
boto3==1.34.0
minio==7.2.0

# Data Lake & Warehouse
delta-lake==0.15.0
apache-airflow==2.7.3
great-expectations==0.18.5

# Scheduling & Monitoring
schedule==1.2.0
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# Image Processing (for menu images)
pillow==10.1.0
opencv-python==4.8.1.78

# Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# API & HTTP
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
aiohttp==3.9.1

# Configuration & Validation
pyyaml==6.0.1
marshmallow==3.20.1
cerberus==1.3.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0